api/requirements.txt
beautifulsoup4
lemminflect
Cython
srsly==1.0.2
cymem==2.0.3
scipy
spacy==2.3.2
tqdm
grobid-quantities-client
fastapi
fastapi-mail
uvicorn
python-multipart
pandas==1.0.5
pip==20.1.1
boto3==1.9.66
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
nltk==3.5
requests-aws4auth==1.0
tensorflow==2.4.1
transformers==4.1.1
sentencepiece==0.1.91